Create Billing-System.py
This is a simple GUI based application which is very easy to understand and use. It uses the Tkinter module for the GUI. The user can address the order and view the total receipt of the items which displays receipt number and number of their food/drinks items with the total amount. There’s also an extra calculator feature for the users.